requirements.txt
# Core Web Framework
shiny>=0.10.0
shinywidgets>=0.3.0

# Machine Learning & NLP
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0
spacy>=3.7.0
scikit-learn>=1.3.0
scipy>=1.10.0
captum>=0.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Utilities
Pillow>=10.0.0
nltk>=3.8.1

# API & Datasets
google-generativeai>=0.3.0
datasets>=2.14.0
